+++ XADMaster/XADGzipParser.m
@@ -13,7 +13,7 @@
 +(BOOL)recognizeFileWithHandle:(CSHandle *)handle firstBytes:(NSData *)data name:(NSString *)name;
 {
 	const uint8_t *bytes=[data bytes];
-	int length=[data length];
+	NSInteger length=[data length];
 
 	if(length<3) return NO;
 	if(bytes[0]==0x1f)
@@ -120,7 +120,7 @@
 name:(NSString *)name propertiesToAdd:(NSMutableDictionary *)props
 {
 	const uint8_t *bytes=[data bytes];
-	int length=[data length];
+	NSInteger length=[data length];
 
 	if(length<5) return NO;
 	if(length>5000) length=5000;
